Fixed many problems in sending messages
A prior commit had changed the files array to a two-level array $file[fileid][tailinfo,productname]
This broke the File::Tail interface that was passed a reference to @file
so this change was reverted, the array @names was created to hold the product names, and @pending was created to hold the prior data for multi-line logs

many debug prints were put in, but they are all commented out.

In addition, loglog will now output any pending data when a new timeslice rolls around and there is no new data to send.